Rename Token.getToken() ~> CSRF.getToken()
